chore: move keyboard nav and FF wrapper
