+++ src/data-workspace/data-entry-cell/index.js
@@ -1,4 +1,3 @@
-export { FinalFormWrapper } from './final-form-wrapper.js'
 export { DataEntryCell } from './data-entry-cell.js'
 export { DATA_VALUE_MUTATION_KEY } from './use-data-value-mutation.js'
 export { useActiveCell } from './use-active-cell.js'
