+++ src/data-workspace/data-workspace.js
@@ -2,10 +2,10 @@ import i18n from '@dhis2/d2-i18n'
 import { CircularLoader, NoticeBox } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { FinalFormWrapper } from './data-entry-cell/index.js'
-import { KeyboardNavManager } from './data-entry-cell/keyboard-nav-manager.js'
 import styles from './data-workspace.module.css'
 import { EntryForm } from './entry-form.js'
+import { FinalFormWrapper } from './final-form-wrapper.js'
+import { KeyboardNavManager } from './keyboard-nav-manager.js'
 import { useDataSet } from './use-data-set.js'
 import { useInitialDataValues } from './use-initial-data-values.js'
